Added the primitive CloudSQL support. Extracted SQLHub from MySQL. Since SQLHub doesn't depend on MySQLdb, it can also be used with Google's Cloud SQL API. Because escapeString, __connect, and __tryConnect are now implemented by subclasses, they are replaced by stubs with raise NotImplemented and the latter two were renamed to connect and __tryConnect respectively. Also renamed __connection to SQLHub.connection.
